consumer/base_consumer.py
from kafka import KafkaConsumer
import psycopg2
import os
import json

def connect_kafka(topic, group_id):
    bootstrap_servers = os.getenv('KAFKA_BOOTSTRAP_SERVERS', 'localhost:9092')
    consumer = KafkaConsumer(
        topic,
        bootstrap_servers=bootstrap_servers,
        group_id=group_id,
        value_deserializer=lambda x: json.loads(x.decode('utf-8')),
        enable_auto_commit=True,
        auto_offset_reset='latest',
    )
    return consumer
